+++ src/platform/console-clean.ts
@@ -16,4 +16,3 @@ console.warn = (...args) => {
   }
   _warn(...args);
 };
-